requirements.txt
alembic==1.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
docx2txt==0.8
dominate==2.5.1
email-validator==1.1.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-JWT-Extended==3.24.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
geographiclib==1.50
geopy==2.0.0
get==0.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
Mako==1.1.3
MarkupSafe==1.1.1
nltk==3.5
numpy==1.19.0
pdfkit==0.6.1
post==0.0.0
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
query-string==0.0.0
regex==2020.6.8
request==0.0.0
requests==2.24.0
scikit-learn==0.23.1
scipy==1.5.1
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.18
tabula==1.0.5
threadpoolctl==2.1.0
tika==1.24
tqdm==4.47.0
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
